requirements.txt
geojson
pandas
pip
python
python-dotenv
psycopg2
sqlalchemy
docker-compose
make
python-dotenv